Move common filtering logic from trigger to pjutil.
