+++ prow/pjutil/BUILD.bazel
@@ -11,6 +11,7 @@ load(
 go_library(
     name = "go_default_library",
     srcs = [
+        "filter.go",
         "pjutil.go",
         "pprof.go",
         "tot.go",
@@ -46,6 +47,7 @@ filegroup(
 go_test(
     name = "go_default_test",
     srcs = [
+        "filter_test.go",
         "pjutil_test.go",
         "tot_test.go",
     ],
