+++ prow/pjutil/filter.go
@@ -1,0 +1,61 @@
+/*
+Copyright 2019 The Kubernetes Authors.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package pjutil
+
+import (
+	"regexp"
+
+	"k8s.io/test-infra/prow/config"
+)
+
+var TestAllRe = regexp.MustCompile(`(?m)^/test all,?($|\s.*)`)
+
+// Filter digests a presubmit config to determine if:
+//  - we can be certain that the presubmit should run
+//  - we know that the presubmit is forced to run
+//  - what the default behavior should be if the presubmit
+//    runs conditionally and does not match trigger conditions
+type Filter func(p config.Presubmit) (shouldRun bool, forcedToRun bool, defaultBehavior bool)
+
+// CommandFilter builds a filter for `/test foo`
+func CommandFilter(body string) Filter {
+	return func(p config.Presubmit) (bool, bool, bool) {
+		return p.TriggerMatches(body), p.TriggerMatches(body), true
+	}
+}
+
+// TestAllFilter builds a filter for the automatic behavior of `/test all`.
+// Jobs that explicitly match `/test all` in their trigger regex will be
+// handled by a commandFilter for the comment in question.
+func TestAllFilter() Filter {
+	return func(p config.Presubmit) (bool, bool, bool) {
+		return !p.NeedsExplicitTrigger(), false, false
+	}
+}
+
+// AggregateFilter builds a filter that evaluates the child filters in order
+// and returns the first match
+func AggregateFilter(filters []Filter) Filter {
+	return func(presubmit config.Presubmit) (bool, bool, bool) {
+		for _, filter := range filters {
+			if shouldRun, forced, defaults := filter(presubmit); shouldRun {
+				return shouldRun, forced, defaults
+			}
+		}
+		return false, false, false
+	}
+}
