+++ prow/plugins/trigger/BUILD.bazel
@@ -22,6 +22,7 @@ go_test(
         "//prow/github:go_default_library",
         "//prow/github/fakegithub:go_default_library",
         "//prow/labels:go_default_library",
+        "//prow/pjutil:go_default_library",
         "//prow/plugins:go_default_library",
         "//vendor/github.com/pkg/errors:go_default_library",
         "//vendor/github.com/sirupsen/logrus:go_default_library",
