+++ prow/plugins/trigger/generic-comment.go
@@ -18,6 +18,7 @@ package trigger
 
 import (
 	"fmt"
+	"k8s.io/test-infra/prow/pjutil"
 	"regexp"
 
 	"github.com/sirupsen/logrus"
@@ -29,7 +30,6 @@ import (
 )
 
 var okToTestRe = regexp.MustCompile(`(?m)^/ok-to-test\s*$`)
-var testAllRe = regexp.MustCompile(`(?m)^/test all,?($|\s.*)`)
 var retestRe = regexp.MustCompile(`(?m)^/retest\s*$`)
 
 func handleGenericComment(c Client, trigger plugins.Trigger, gc github.GenericCommentEvent) error {
@@ -44,7 +44,7 @@ func handleGenericComment(c Client, trigger plugins.Trigger, gc github.GenericCo
 		return nil
 	}
 	// Skip comments not germane to this plugin
-	if !retestRe.MatchString(gc.Body) && !okToTestRe.MatchString(gc.Body) && !testAllRe.MatchString(gc.Body) {
+	if !retestRe.MatchString(gc.Body) && !okToTestRe.MatchString(gc.Body) && !pjutil.TestAllRe.MatchString(gc.Body) {
 		matched := false
 		for _, presubmit := range c.Config.Presubmits[gc.Repo.FullName] {
 			matched = matched || presubmit.TriggerMatches(gc.Body)
@@ -160,7 +160,7 @@ type changesGetter interface {
 
 // filterPresubmits determines which presubmits should run and which should be skipped
 // by evaluating the user-provided filter.
-func filterPresubmits(filter filter, gitHubClient changesGetter, pr *github.PullRequest, presubmits []config.Presubmit, logger *logrus.Entry) ([]config.Presubmit, []config.Presubmit, error) {
+func filterPresubmits(filter pjutil.Filter, gitHubClient changesGetter, pr *github.PullRequest, presubmits []config.Presubmit, logger *logrus.Entry) ([]config.Presubmit, []config.Presubmit, error) {
 	org, repo, number, branch := pr.Base.Repo.Owner.Login, pr.Base.Repo.Name, pr.Number, pr.Base.Ref
 	changes := config.NewGitHubDeferredChangedFilesProvider(gitHubClient, org, repo, number)
 	var toTrigger []config.Presubmit
@@ -209,15 +209,15 @@ type statusGetter interface {
 	GetCombinedStatus(org, repo, ref string) (*github.CombinedStatus, error)
 }
 
-func presubmitFilter(honorOkToTest bool, statusGetter statusGetter, body, org, repo, sha string, logger *logrus.Entry) (filter, error) {
+func presubmitFilter(honorOkToTest bool, statusGetter statusGetter, body, org, repo, sha string, logger *logrus.Entry) (pjutil.Filter, error) {
 	// the filters determine if we should check whether a job should run, whether
 	// it should run regardless of whether its triggering conditions match, and
 	// what the default behavior should be for that check. Multiple filters
 	// can match a single presubmit, so it is important to order them correctly
 	// as they have precedence -- filters that override the false default should
 	// match before others. We order filters by amount of specificity.
-	var filters []filter
-	filters = append(filters, commandFilter(body))
+	var filters []pjutil.Filter
+	filters = append(filters, pjutil.CommandFilter(body))
 	if retestRe.MatchString(body) {
 		logger.Debug("Using retest filter.")
 		combinedStatus, err := statusGetter.GetCombinedStatus(org, repo, sha)
@@ -235,52 +235,16 @@ func presubmitFilter(honorOkToTest bool, statusGetter statusGetter, body, org, r
 
 		filters = append(filters, retestFilter(failedContexts, allContexts))
 	}
-	if (honorOkToTest && okToTestRe.MatchString(body)) || testAllRe.MatchString(body) {
+	if (honorOkToTest && okToTestRe.MatchString(body)) || pjutil.TestAllRe.MatchString(body) {
 		logger.Debug("Using test-all filter.")
-		filters = append(filters, testAllFilter())
+		filters = append(filters, pjutil.TestAllFilter())
 	}
-	return aggregateFilter(filters), nil
-}
-
-// filter digests a presubmit config to determine if:
-//  - we can be certain that the presubmit should run
-//  - we know that the presubmit is forced to run
-//  - what the default behavior should be if the presubmit
-//    runs conditionally and does not match trigger conditions
-type filter func(p config.Presubmit) (shouldRun bool, forcedToRun bool, defaultBehavior bool)
-
-// commandFilter builds a filter for `/test foo`
-func commandFilter(body string) filter {
-	return func(p config.Presubmit) (bool, bool, bool) {
-		return p.TriggerMatches(body), p.TriggerMatches(body), true
-	}
+	return pjutil.AggregateFilter(filters), nil
 }
 
 // retestFilter builds a filter for `/retest`
-func retestFilter(failedContexts, allContexts sets.String) filter {
+func retestFilter(failedContexts, allContexts sets.String) pjutil.Filter {
 	return func(p config.Presubmit) (bool, bool, bool) {
 		return failedContexts.Has(p.Context) || (!p.NeedsExplicitTrigger() && !allContexts.Has(p.Context)), false, true
-	}
-}
-
-// testAllFilter builds a filter for the automatic behavior of `/test all`.
-// Jobs that explicitly match `/test all` in their trigger regex will be
-// handled by a commandFilter for the comment in question.
-func testAllFilter() filter {
-	return func(p config.Presubmit) (bool, bool, bool) {
-		return !p.NeedsExplicitTrigger(), false, false
-	}
-}
-
-// aggregateFilter builds a filter that evaluates the child filters in order
-// and returns the first match
-func aggregateFilter(filters []filter) filter {
-	return func(presubmit config.Presubmit) (bool, bool, bool) {
-		for _, filter := range filters {
-			if shouldRun, forced, defaults := filter(presubmit); shouldRun {
-				return shouldRun, forced, defaults
-			}
-		}
-		return false, false, false
 	}
 }
