+++ prow/plugins/trigger/generic-comment_test.go
@@ -18,6 +18,7 @@ package trigger
 
 import (
 	"fmt"
+	"k8s.io/test-infra/prow/pjutil"
 	"log"
 	"reflect"
 	"testing"
@@ -1367,62 +1368,6 @@ func TestPresubmitFilter(t *testing.T) {
 	}
 }
 
-func TestCommandFilter(t *testing.T) {
-	var testCases = []struct {
-		name       string
-		body       string
-		presubmits []config.Presubmit
-		expected   [][]bool
-	}{
-		{
-			name: "command filter matches jobs whose triggers match the body",
-			body: "/test trigger",
-			presubmits: []config.Presubmit{
-				{
-					JobBase: config.JobBase{
-						Name: "trigger",
-					},
-					Trigger:      `(?m)^/test (?:.*? )?trigger(?: .*?)?$`,
-					RerunCommand: "/test trigger",
-				},
-				{
-					JobBase: config.JobBase{
-						Name: "other-trigger",
-					},
-					Trigger:      `(?m)^/test (?:.*? )?other-trigger(?: .*?)?$`,
-					RerunCommand: "/test other-trigger",
-				},
-			},
-			expected: [][]bool{{true, true, true}, {false, false, true}},
-		},
-	}
-
-	for _, testCase := range testCases {
-		t.Run(testCase.name, func(t *testing.T) {
-			if len(testCase.presubmits) != len(testCase.expected) {
-				t.Fatalf("%s: have %d presubmits but only %d expected filter outputs", testCase.name, len(testCase.presubmits), len(testCase.expected))
-			}
-			if err := config.SetPresubmitRegexes(testCase.presubmits); err != nil {
-				t.Fatalf("%s: could not set presubmit regexes: %v", testCase.name, err)
-			}
-			filter := commandFilter(testCase.body)
-			for i, presubmit := range testCase.presubmits {
-				actualFiltered, actualForced, actualDefault := filter(presubmit)
-				expectedFiltered, expectedForced, expectedDefault := testCase.expected[i][0], testCase.expected[i][1], testCase.expected[i][2]
-				if actualFiltered != expectedFiltered {
-					t.Errorf("%s: filter did not evaluate correctly, expected %v but got %v for %v", testCase.name, expectedFiltered, actualFiltered, presubmit.Name)
-				}
-				if actualForced != expectedForced {
-					t.Errorf("%s: filter did not determine forced correctly, expected %v but got %v for %v", testCase.name, expectedForced, actualForced, presubmit.Name)
-				}
-				if actualDefault != expectedDefault {
-					t.Errorf("%s: filter did not determine default correctly, expected %v but got %v for %v", testCase.name, expectedDefault, actualDefault, presubmit.Name)
-				}
-			}
-		})
-	}
-}
-
 func TestRetestFilter(t *testing.T) {
 	var testCases = []struct {
 		name           string
@@ -1508,81 +1453,6 @@ func TestRetestFilter(t *testing.T) {
 	}
 }
 
-func TestTestAllFilter(t *testing.T) {
-	var testCases = []struct {
-		name       string
-		presubmits []config.Presubmit
-		expected   [][]bool
-	}{
-		{
-			name: "test all filter matches jobs which do not require human triggering",
-			presubmits: []config.Presubmit{
-				{
-					JobBase: config.JobBase{
-						Name: "always-runs",
-					},
-					AlwaysRun: true,
-				},
-				{
-					JobBase: config.JobBase{
-						Name: "runs-if-changed",
-					},
-					AlwaysRun: false,
-					RegexpChangeMatcher: config.RegexpChangeMatcher{
-						RunIfChanged: "sometimes",
-					},
-				},
-				{
-					JobBase: config.JobBase{
-						Name: "runs-if-triggered",
-					},
-					Reporter: config.Reporter{
-						Context: "runs-if-triggered",
-					},
-					Trigger:      `(?m)^/test (?:.*? )?trigger(?: .*?)?$`,
-					RerunCommand: "/test trigger",
-				},
-				{
-					JobBase: config.JobBase{
-						Name: "literal-test-all-trigger",
-					},
-					Reporter: config.Reporter{
-						Context: "runs-if-triggered",
-					},
-					Trigger:      `(?m)^/test (?:.*? )?all(?: .*?)?$`,
-					RerunCommand: "/test all",
-				},
-			},
-			expected: [][]bool{{true, false, false}, {true, false, false}, {false, false, false}, {false, false, false}},
-		},
-	}
-
-	for _, testCase := range testCases {
-		t.Run(testCase.name, func(t *testing.T) {
-			if len(testCase.presubmits) != len(testCase.expected) {
-				t.Fatalf("%s: have %d presubmits but only %d expected filter outputs", testCase.name, len(testCase.presubmits), len(testCase.expected))
-			}
-			if err := config.SetPresubmitRegexes(testCase.presubmits); err != nil {
-				t.Fatalf("%s: could not set presubmit regexes: %v", testCase.name, err)
-			}
-			filter := testAllFilter()
-			for i, presubmit := range testCase.presubmits {
-				actualFiltered, actualForced, actualDefault := filter(presubmit)
-				expectedFiltered, expectedForced, expectedDefault := testCase.expected[i][0], testCase.expected[i][1], testCase.expected[i][2]
-				if actualFiltered != expectedFiltered {
-					t.Errorf("%s: filter did not evaluate correctly, expected %v but got %v for %v", testCase.name, expectedFiltered, actualFiltered, presubmit.Name)
-				}
-				if actualForced != expectedForced {
-					t.Errorf("%s: filter did not determine forced correctly, expected %v but got %v for %v", testCase.name, expectedForced, actualForced, presubmit.Name)
-				}
-				if actualDefault != expectedDefault {
-					t.Errorf("%s: filter did not determine default correctly, expected %v but got %v for %v", testCase.name, expectedDefault, actualDefault, presubmit.Name)
-				}
-			}
-		})
-	}
-}
-
 func TestDetermineSkippedPresubmits(t *testing.T) {
 	var testCases = []struct {
 		name                      string
@@ -1643,7 +1513,7 @@ func (c *fakeChangesGetter) GetPullRequestChanges(org, repo string, number int) 
 func TestFilterPresubmits(t *testing.T) {
 	var testCases = []struct {
 		name                              string
-		filter                            filter
+		filter                            pjutil.Filter
 		presubmits                        []config.Presubmit
 		changesError                      bool
 		expectedToTrigger, expectedToSkip []config.Presubmit
